+++ packages/sit-onyx/src/components/OnyxSelect/types.ts
@@ -64,7 +64,8 @@ export type OnyxSelectProps<TValue extends SelectOptionValue = SelectOptionValue
   SelectModelValueProps<TValue> &
   SelectSearchProps &
   Omit<OnyxSelectInputProps<TValue>, "density" | "modelValue"> &
-  AutofocusProp & {
+  AutofocusProp &
+  Pick<BaseSelectOption, "truncation"> & {
     /**
      * If true, the select popover is expanded and visible.
      * Property is managed internally, when undefined.
@@ -88,7 +89,7 @@ export type OnyxSelectProps<TValue extends SelectOptionValue = SelectOptionValue
 
 export type SelectOption<TValue extends SelectOptionValue = SelectOptionValue> = Pick<
   BaseSelectOption<TValue>,
-  "value" | "label" | "disabled"
+  "value" | "label" | "disabled" | "truncation"
 > &
   Pick<OnyxSelectOptionProps, "icon"> & {
     /**