+++ packages/sit-onyx/src/types/themes.ts
@@ -1,0 +1,1 @@
+export type OnyxTheme = "kaufland" | "lidl" | "onyx" | "twogo";