+++ packages/sit-onyx/CHANGELOG.md
@@ -1,5 +1,17 @@
 # sit-onyx
 
+## 1.0.0-beta.3
+
+### Minor Changes
+
+- 67a5e50: feat(OnyxSelect, OnyxRadioGroup, OnyxCheckboxGroup): Implement truncation on higher level
+
+## 1.0.0-beta.2
+
+### Minor Changes
+
+- 15114f7: feat(OnyxSelect): Implement truncation for OnyxSelectOptions
+
 ## 1.0.0-beta.1
 
 ### Major Changes